requirements.txt
flask==3.0.2
pycryptodome==3.20.0
python-dotenv==1.0.1
werkzeug==3.0.1